Corrige apresentação de valor de campo que foi confundido com conteúdo XML
