+++ src/scielo/bin/cfg/xpm_version.txt
@@ -1,4 +1,12 @@
-﻿4.0.092 - 21/11/2016
+﻿4.0.092 - 21/11/2016 Release 12/12/2016
+=======================================
+- XPM
+  - (bug) [#2451, #2440, #2436] Fix images report
+  - (bug) [#2454] Fix missing preview of equation
+  - (bug) [#2302] Invalid error about missing xref (aff) em sub-article
+
+
+4.0.092 - 21/11/2016
 ====================
 - General:
   - (new) [#2287] Solve HTTP Error Proxy Authentication Required
